Separate features from examples stories
Also removes the component from the story meta - this shows the props of
AnimationProvider in the controls panel which cannot be edited in example
or feature stories and may cause confusion.
